refactor: inline shared config packages and update dependencies
- Eliminate packages/eslint-config and packages/tsconfig
- Inline tsconfig base into root, extend from each project
- Inline eslint base config into root, import from each project
- Update all dependencies to latest (TS 6.0, ESLint 10, Next.js 16)
- Replace eslint-plugin-import with eslint-plugin-import-x (ESLint 10 compat)
- Replace eslintConfigPrettier.rules with eslintConfigPrettier (flat config)
- Use @next/eslint-plugin-next flat config directly instead of FlatCompat
- Remove deprecated TS options (downlevelIteration, alwaysStrict)
- Add rootDir to server tsconfig
- Switch web lint from 'next lint' to 'eslint .'
