+++ apps/web/package.json
@@ -7,25 +7,26 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "eslint ."
   },
   "dependencies": {
     "@monorepo/types": "workspace:*",
     "@monorepo/ui": "workspace:*",
     "@monorepo/utils": "workspace:*",
     "class-variance-authority": "^0.7.1",
-    "lucide-react": "^0.525.0",
-    "next": "^15.4.5",
+    "lucide-react": "^1.8.0",
+    "next": "^16.2.3",
     "next-themes": "^0.4.6",
-    "react": "^19.1.1",
-    "react-dom": "^19.1.1"
+    "react": "^19.2.5",
+    "react-dom": "^19.2.5"
   },
   "devDependencies": {
-    "@monorepo/eslint-config": "workspace:*",
-    "@monorepo/tsconfig": "workspace:*",
-    "@tailwindcss/postcss": "^4.1.12",
-    "@types/react": "^19.1.9",
-    "@types/react-dom": "^19.1.7",
-    "tailwindcss": "^4.1.12"
+    "@next/eslint-plugin-next": "^16.2.3",
+    "@tailwindcss/postcss": "^4.2.2",
+    "@types/react": "^19.2.14",
+    "@types/react-dom": "^19.2.3",
+    "globals": "^17.5.0",
+    "tailwindcss": "^4.2.2",
+    "typescript-eslint": "^8.58.2"
   }
 }
