--- packages/eslint-config/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "@monorepo/eslint-config",
-  "version": "0.0.1",
-  "private": true,
-  "main": "./base.js",
-  "type": "module",
-  "exports": {
-    "./base": "./base.js",
-    "./next": "./next.js",
-    "./express": "./express.js",
-    "./react-internal": "./react-internal.js"
-  },
-  "files": [
-    "next.js",
-    "express.js",
-    "base.js",
-    "react-internal.js"
-  ],
-  "devDependencies": {
-    "@eslint/compat": "^1.3.2",
-    "@eslint/eslintrc": "^3.3.1",
-    "@eslint/js": "^9.34.0",
-    "@next/eslint-plugin-next": "^15.5.2",
-    "@typescript-eslint/eslint-plugin": "^8.41.0",
-    "@typescript-eslint/parser": "^8.41.0",
-    "eslint-config-prettier": "^10.1.8",
-    "eslint-plugin-import": "^2.32.0",
-    "eslint-plugin-only-warn": "^1.1.0",
-    "eslint-plugin-perfectionist": "^4.15.0",
-    "eslint-plugin-prettier": "^5.5.4",
-    "eslint-plugin-security": "^3.0.1",
-    "globals": "^16.3.0",
-    "typescript-eslint": "^8.41.0"
-  }
-}
