--- packages/eslint-config/react-internal.js
@@ -1,34 +1,0 @@
-import globals from "globals";
-import tseslint from "typescript-eslint";
-
-import baseConfig from "./base.js";
-
-/** @type {import("eslint").Linter.Config[]} */
-export default [
-  ...baseConfig,
-  {
-    languageOptions: {
-      globals: {
-        ...globals.browser,
-        JSX: true,
-        React: true,
-      },
-      parserOptions: {
-        ecmaVersion: "latest",
-        sourceType: "module",
-      },
-    },
-    plugins: {
-      "@typescript-eslint": tseslint.plugin,
-    },
-    rules: {
-      "react/prop-types": "off",
-      "react/react-in-jsx-scope": "off",
-    },
-    settings: {
-      react: {
-        version: "detect",
-      },
-    },
-  },
-];
