--- packages/tsconfig/express.json
@@ -1,22 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "display": "ExpressJS Server",
-  "ts-node": {
-    "compilerOptions": {
-      "module": "commonjs",
-      "moduleResolution": "Node10"
-    }
-  },
-  "compilerOptions": {
-    "target": "ES2022",
-    "verbatimModuleSyntax": false,
-    "module": "NodeNext",
-    "sourceMap": true,
-    "moduleResolution": "nodenext",
-    "resolveJsonModule": true,
-    "esModuleInterop": true,
-    "noImplicitAny": true,
-    "removeComments": true,
-    "strict": true
-  }
-}
