--- packages/tsconfig/package.json
@@ -1,10 +1,0 @@
-{
-  "name": "@monorepo/tsconfig",
-  "author": "Ives Furtado | @ivesfurtado",
-  "version": "0.0.0",
-  "private": true,
-  "license": "MIT",
-  "publishConfig": {
-    "access": "public"
-  }
-}
