--- packages/tsconfig/types.json
@@ -1,10 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "display": "Shared Types",
-  "extends": "./base.json",
-  "compilerOptions": {
-    "declaration": true,
-    "emitDecoratorMetadata": true,
-    "experimentalDecorators": true,
-  },
-}
