+++ packages/types/tsconfig.json
@@ -1,10 +1,10 @@
 {
-  "extends": "@monorepo/tsconfig/types.json",
-  "include": [
-      "**/*.ts",
-      "eslint.config.mjs"
-  ],
-  "exclude": [
-    "node_modules"
-  ],
+  "extends": "../../tsconfig.json",
+  "compilerOptions": {
+    "declaration": true,
+    "emitDecoratorMetadata": true,
+    "experimentalDecorators": true
+  },
+  "include": ["**/*.ts"],
+  "exclude": ["node_modules"]
 }
