+++ packages/ui/eslint.config.mjs
@@ -1,4 +1,34 @@
-import reactInternalConfig from "@monorepo/eslint-config/react-internal";
+import globals from "globals";
+import tseslint from "typescript-eslint";
+
+import baseConfig from "../../eslint.config.mjs";
 
 /** @type {import("eslint").Linter.Config[]} */
-export default [...reactInternalConfig];
+export default [
+  ...baseConfig,
+  {
+    languageOptions: {
+      globals: {
+        ...globals.browser,
+        JSX: true,
+        React: true,
+      },
+      parserOptions: {
+        ecmaVersion: "latest",
+        sourceType: "module",
+      },
+    },
+    plugins: {
+      "@typescript-eslint": tseslint.plugin,
+    },
+    rules: {
+      "react/prop-types": "off",
+      "react/react-in-jsx-scope": "off",
+    },
+    settings: {
+      react: {
+        version: "detect",
+      },
+    },
+  },
+];
