+++ packages/ui/tsconfig.json
@@ -1,18 +1,22 @@
 {
-  "extends": "@monorepo/tsconfig/ui.json",
-  // https://stackoverflow.com/questions/58337089/extending-tsconfig-json-file-doesnt-seem-to-extend-anything
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
+    "allowJs": true,
+    "skipLibCheck": true,
+    "strict": false,
+    "esModuleInterop": true,
+    "forceConsistentCasingInFileNames": true,
+    "noEmit": true,
+    "incremental": true,
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "jsx": "preserve",
     "baseUrl": ".",
     "paths": {
       "@/*": ["src/*"],
       "@monorepo/ui/*": ["./src/*"]
     }
   },
-  "include": [
-    "src/**/*.ts",
-    "src/**/*.tsx",
-    "tailwind.config.ts",
-    "eslint.config.mjs"
-  ],
+  "include": ["src/**/*.ts", "src/**/*.tsx", "tailwind.config.ts"],
   "exclude": ["node_modules", "dist"]
 }
