+++ packages/utils/package.json
@@ -12,11 +12,7 @@
   },
   "dependencies": {
     "clsx": "^2.1.1",
-    "tailwind-merge": "^3.3.1"
-  },
-  "devDependencies": {
-    "@monorepo/eslint-config": "workspace:*",
-    "@monorepo/tsconfig": "workspace:*"
+    "tailwind-merge": "^3.5.0"
   },
   "exports": {
     ".": "./src",
