+++ packages/utils/tsconfig.json
@@ -1,10 +1,20 @@
 {
-  "extends": "@monorepo/tsconfig/utils.json",
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
+    "allowJs": true,
+    "skipLibCheck": true,
+    "strict": false,
+    "forceConsistentCasingInFileNames": true,
+    "noEmit": true,
+    "incremental": true,
+    "esModuleInterop": true,
+    "resolveJsonModule": true,
+    "isolatedModules": true,
+    "jsx": "preserve",
     "paths": {
       "@/*": ["./*"]
     }
   },
-  "include": ["src/**/*.ts", "eslint.config.mjs"],
+  "include": ["src/**/*.ts"],
   "exclude": ["node_modules"]
 }
